Change int to long + bugfix select button  (#1092)
* replacing fake type authType with ldap

* code cleanup (todos)

* cleaning up demo user roles

* documented sample data removal need on start page

* almost working
load template does not update "select all" button yet

* adding select all parameter to certificatino

* Change int to long + bugfix select button

closes #923

bugfix select button

* fixes

+++ roles/lib/files/FWO_API_Client/Data/Management.cs
@@ -83,7 +83,7 @@ namespace FWO.Api.Data
 
         // [JsonPropertyName("pointInTime")]
         // public DateTime ReportTime { get; set; }
-        public int? RelevantImportId { get; set; }
+        public long? RelevantImportId { get; set; }
 
         //[JsonPropertyName("rule_id")]
         public List<long> ReportedRuleIds { get; set; }
